[INPLACE-285] refactor: security module 추가
