+++ backend/inplace-security/src/main/java/config/JwtProperties.java
@@ -1,0 +1,13 @@
+package config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "spring.jwt")
+public record JwtProperties(
+        String secret,
+        Long accessTokenExpiredTime,
+        Long refreshTokenExpiredTime,
+        Long adminAccessTokenExpiredTime
+) {
+
+}
