+++ backend/inplace-security/src/main/java/config/SecurityFilterConfig.java
@@ -1,0 +1,31 @@
+package config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.micrometer.core.instrument.MeterRegistry;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import filter.ApiExecutionMetricsFilter;
+import filter.AuthorizationFilter;
+import filter.ExceptionHandlingFilter;
+import util.JwtUtil;
+
+@Configuration
+public class SecurityFilterConfig {
+
+    @Bean
+    public AuthorizationFilter authorizationFilter(JwtUtil jwtUtil) {
+        return new AuthorizationFilter(jwtUtil);
+    }
+
+    @Bean
+    public ExceptionHandlingFilter exceptionHandlingFilter(ObjectMapper objectMapper) {
+        return new ExceptionHandlingFilter(objectMapper);
+    }
+
+    @Bean
+    public ApiExecutionMetricsFilter apiExecutionMetricsFilter(
+        MeterRegistry meterRegistry, ObjectMapper objectMapper
+    ) {
+        return new ApiExecutionMetricsFilter(meterRegistry, objectMapper);
+    }
+}
