+++ backend/inplace-security/src/main/java/filter/TokenType.java
@@ -1,0 +1,14 @@
+package filter;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum TokenType {
+    ACCESS_TOKEN("access_token"),
+    REFRESH_TOKEN("refresh_token");
+
+    private final String value;
+
+}
