+++ backend/inplace-security/src/main/java/util/CookieUtil.java
@@ -1,0 +1,28 @@
+package util;
+
+import org.springframework.http.ResponseCookie;
+
+public class CookieUtil {
+
+    public static ResponseCookie createHttpOnlyCookie(String key, String value, String domain) {
+        return ResponseCookie.from(key, value)
+            .sameSite("None")
+            .secure(true)
+            .path("/")
+            .httpOnly(true)
+            .domain(domain)
+            .maxAge(60 * 60)
+            .build();
+    }
+
+    public static ResponseCookie createCookie(String key, String value, String domain) {
+        return ResponseCookie.from(key, value)
+            .sameSite("None")
+            .secure(true)
+            .path("/")
+            .httpOnly(false)
+            .domain(domain)
+            .maxAge(60 * 60)
+            .build();
+    }
+}
