--- backend/src/main/java/team7/inplace/security/application/CurrentUserProvider.java
@@ -1,23 +1,0 @@
-package team7.inplace.security.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import team7.inplace.global.exception.InplaceException;
-import team7.inplace.global.exception.code.UserErrorCode;
-import team7.inplace.security.util.AuthorizationUtil;
-import user.User;
-import user.jpa.UserJpaRepository;
-
-@Component
-@RequiredArgsConstructor
-public class CurrentUserProvider {
-
-    private final UserJpaRepository userJpaRepository;
-
-    public User getCurrentUser() {
-        Long userId = AuthorizationUtil.getUserIdOrThrow();
-
-        return userJpaRepository.findById(userId)
-            .orElseThrow(() -> InplaceException.of(UserErrorCode.NOT_FOUND));
-    }
-}
