--- backend/src/main/java/team7/inplace/security/config/JwtProperties.java
@@ -1,13 +1,0 @@
-package team7.inplace.security.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "spring.jwt")
-public record JwtProperties(
-        String secret,
-        Long accessTokenExpiredTime,
-        Long refreshTokenExpiredTime,
-        Long adminAccessTokenExpiredTime
-) {
-
-}
