--- backend/src/main/java/team7/inplace/security/config/SecurityHandlerConfig.java
@@ -1,32 +1,0 @@
-package team7.inplace.security.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import team7.inplace.security.handler.CustomAccessDeniedHandler;
-import team7.inplace.security.handler.CustomFailureHandler;
-import team7.inplace.security.handler.CustomSuccessHandler;
-import team7.inplace.security.util.JwtUtil;
-import team7.inplace.token.application.RefreshTokenService;
-
-@Configuration
-public class SecurityHandlerConfig {
-
-    @Bean
-    public CustomSuccessHandler customSuccessHandler(
-            JwtUtil jwtUtil,
-            RefreshTokenService refreshTokenService
-    ) {
-        return new CustomSuccessHandler(jwtUtil, refreshTokenService);
-    }
-
-    @Bean
-    public CustomFailureHandler customFailureHandler(ObjectMapper objectMapper) {
-        return new CustomFailureHandler(objectMapper);
-    }
-
-    @Bean
-    public CustomAccessDeniedHandler customAccessDeniedHandler() {
-        return new CustomAccessDeniedHandler();
-    }
-}
