--- backend/src/main/java/team7/inplace/security/config/SecurityServiceConfig.java
@@ -1,34 +1,0 @@
-package team7.inplace.security.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import team7.inplace.admin.user.application.AdminUserService;
-import team7.inplace.security.application.CustomOAuth2UserService;
-import team7.inplace.security.application.CustomUserDetailsService;
-import team7.inplace.token.application.OauthTokenService;
-import team7.inplace.user.application.UserService;
-
-@Configuration
-public class SecurityServiceConfig {
-
-    @Bean
-    public DefaultOAuth2UserService defaultOAuth2UserService() {
-        return new DefaultOAuth2UserService();
-    }
-
-    @Bean
-    public CustomOAuth2UserService customOAuth2UserService(
-        DefaultOAuth2UserService defaultOAuth2UserService,
-        UserService userService,
-        OauthTokenService oauthTokenService
-    ) {
-        return new CustomOAuth2UserService(defaultOAuth2UserService, userService,
-            oauthTokenService);
-    }
-
-    @Bean
-    public CustomUserDetailsService customUserDetailsService(AdminUserService adminUserService) {
-        return new CustomUserDetailsService(adminUserService);
-    }
-}
