--- backend/src/main/java/team7/inplace/security/config/SecurityUtilConfig.java
@@ -1,58 +1,0 @@
-package team7.inplace.security.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.ProviderManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import team7.inplace.security.application.CustomUserDetailsService;
-import team7.inplace.security.util.JwtUtil;
-import team7.inplace.security.util.TokenEncryptionUtil;
-
-@Configuration
-public class SecurityUtilConfig {
-
-    @Value("${spring.oauth.password}")
-    private String oauthPassword;
-
-    @Value("${spring.oauth.salt}")
-    private String oauthSalt;
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.registerModule(new JavaTimeModule());
-        return mapper;
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-    @Bean
-    public AuthenticationManager authenticationManager(
-        CustomUserDetailsService customUserDetailsService,
-        PasswordEncoder passwordEncoder
-    ) {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(customUserDetailsService);
-        authProvider.setPasswordEncoder(passwordEncoder);
-
-        return new ProviderManager(authProvider);
-    }
-
-    @Bean
-    public JwtUtil jwtUtil(JwtProperties jwtProperties) {
-        return new JwtUtil(jwtProperties);
-    }
-
-    @Bean
-    public TokenEncryptionUtil tokenEncryptionUtil() {
-        return new TokenEncryptionUtil(oauthPassword, oauthSalt);
-    }
-}
