--- backend/src/main/java/team7/inplace/security/filter/TokenType.java
@@ -1,14 +1,0 @@
-package team7.inplace.security.filter;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum TokenType {
-    ACCESS_TOKEN("access_token"),
-    REFRESH_TOKEN("refresh_token");
-
-    private final String value;
-
-}
